Refactored BasicServiceBot into BaseBot
