+++ webofneeds/won-bot/README.md
@@ -8,7 +8,7 @@ Bots in the Web of Needs are generally reactive. Each bot uses [event listeners]
 
 The java application that runs a bot is usually a spring-boot application that loads all the necessary config and then starts the bot. For examples, look at the existing [Bot Apps](src/main/java/won/bot/app).
 
-The [base bot](src/main/java/won/bot/framework/bot/base/) consists of several interfaces and abstract classes that build upon each other. The `Bot` interface extends the `OwnerCallback` interface and is partially implemented in the abstract `BaseBot` class, which describes the methods needed for a minimal viable bot. This is extended by the `BasicServiceBot`, which adds the functionality needed to connect to a node and interact with atoms. The next layer are the `ScheduledActionBot` and the `TriggeredBot`, which add Spring task scheduling with [triggers](https://docs.spring.io/spring/docs/current/javadoc-api/org/springframework/scheduling/Trigger.html). This is then extended by the `EventBot`, which adds the event listener structure described here.
+The [base bot](src/main/java/won/bot/framework/bot/base/) consists of several interfaces and abstract classes that build upon each other. The `Bot` interface extends the `OwnerCallback` interface and is partially implemented in the abstract `BaseBot` class, which describes the methods needed for creating a minimal bot that is able to interact with atoms. This is extended by the `ScheduledActionBot` and the `TriggeredBot`, which add Spring task scheduling with [triggers](https://docs.spring.io/spring/docs/current/javadoc-api/org/springframework/scheduling/Trigger.html). This is then extended by the `EventBot`, which adds the event listener structure described here.
 
 Furthermore, the `FactoryBot` builds upon the `EventBot` and adds some additional functionality for creating `FactoryAtoms`. This results in a different structure in which a matching query can be used to filter which atoms are sent to the bot by the matcher. It also creates an additional atom representing the bot that may be visible to other users. See the [TaxiBot](https://github.com/researchstudio-sat/won-transport/blob/master/won-taxi-bot/src/main/java/won/transport/taxi/bot/impl/TaxiBot.java) for a possible use for `FactoryBot`.
