+++ webofneeds/won-bot/src/main/java/won/bot/framework/bot/base/EventBot.java
@@ -90,7 +90,7 @@ public abstract class EventBot extends TriggeredBot {
             eventBus.publish(new ActEvent());
         } else {
             logger.info("not publishing event for call to act() as the bot is not in state {} but {}",
-                    BotLifecyclePhase.ACTIVE, getLifecyclePhase());
+                            BotLifecyclePhase.ACTIVE, getLifecyclePhase());
         }
     }
 
@@ -100,8 +100,8 @@ public abstract class EventBot extends TriggeredBot {
             eventBus.publish(new MessageFromOtherAtomEvent(con, wonMessage));
         } else {
             logger.info(
-                    "not publishing event for call to onMessageFromOtherAtom() as the bot is not in state {} but {}",
-                    BotLifecyclePhase.ACTIVE, getLifecyclePhase());
+                            "not publishing event for call to onMessageFromOtherAtom() as the bot is not in state {} but {}",
+                            BotLifecyclePhase.ACTIVE, getLifecyclePhase());
         }
     }
 
@@ -111,7 +111,7 @@ public abstract class EventBot extends TriggeredBot {
             eventBus.publish(new AtomHintFromMatcherEvent(wonMessage));
         } else {
             logger.info("not publishing event for call to onAtomHintFromMatcher() as the bot is not in state {} but {}",
-                    BotLifecyclePhase.ACTIVE, getLifecyclePhase());
+                            BotLifecyclePhase.ACTIVE, getLifecyclePhase());
         }
     }
 
@@ -121,8 +121,8 @@ public abstract class EventBot extends TriggeredBot {
             eventBus.publish(new SocketHintFromMatcherEvent(wonMessage));
         } else {
             logger.info(
-                    "not publishing event for call to onSocketHintFromMatcher() as the bot is not in state {} but {}",
-                    BotLifecyclePhase.ACTIVE, getLifecyclePhase());
+                            "not publishing event for call to onSocketHintFromMatcher() as the bot is not in state {} but {}",
+                            BotLifecyclePhase.ACTIVE, getLifecyclePhase());
         }
     }
 
@@ -132,7 +132,7 @@ public abstract class EventBot extends TriggeredBot {
             eventBus.publish(new CloseFromOtherAtomEvent(con, wonMessage));
         } else {
             logger.info("not publishing event for call to onClose() as the bot is not in state {} but {}",
-                    BotLifecyclePhase.ACTIVE, getLifecyclePhase());
+                            BotLifecyclePhase.ACTIVE, getLifecyclePhase());
         }
     }
 
@@ -142,7 +142,7 @@ public abstract class EventBot extends TriggeredBot {
             eventBus.publish(new OpenFromOtherAtomEvent(con, wonMessage));
         } else {
             logger.info("not publishing event for call to onOpenFromOtherAtom() as the bot is not in state {} but {}",
-                    BotLifecyclePhase.ACTIVE, getLifecyclePhase());
+                            BotLifecyclePhase.ACTIVE, getLifecyclePhase());
         }
     }
 
@@ -152,31 +152,31 @@ public abstract class EventBot extends TriggeredBot {
             eventBus.publish(new ConnectFromOtherAtomEvent(con, wonMessage));
         } else {
             logger.info(
-                    "not publishing event for call to onConnectFromOtherAtom() as the bot is not in state {} but {}",
-                    BotLifecyclePhase.ACTIVE, getLifecyclePhase());
+                            "not publishing event for call to onConnectFromOtherAtom() as the bot is not in state {} but {}",
+                            BotLifecyclePhase.ACTIVE, getLifecyclePhase());
         }
     }
 
     @Override
     public final void onNewAtomCreated(final URI atomUri, final URI wonNodeUri, final Dataset atomDataset)
-            throws Exception {
+                    throws Exception {
         if (getLifecyclePhase().isActive()) {
             eventBus.publish(new AtomCreatedEvent(atomUri, wonNodeUri, atomDataset, SocketType.ChatSocket));
         } else {
             logger.info("not publishing event for call to onNewAtomCreated() as the bot is not in state {} but {}",
-                    BotLifecyclePhase.ACTIVE, getLifecyclePhase());
+                            BotLifecyclePhase.ACTIVE, getLifecyclePhase());
         }
     }
 
     @Override
     public final void onNewAtomCreatedNotificationForMatcher(final URI wonNodeURI, final URI atomUri,
-            final Dataset wonMessageDataset) {
+                    final Dataset wonMessageDataset) {
         if (getLifecyclePhase().isActive()) {
             Dataset dataset = getEventListenerContext().getLinkedDataSource().getDataForResource(atomUri);
             eventBus.publish(new AtomCreatedEventForMatcher(atomUri, dataset));
         } else {
             logger.info("not publishing event for call to onNewAtomCreated() as the bot is not in state {} but {}",
-                    BotLifecyclePhase.ACTIVE, getLifecyclePhase());
+                            BotLifecyclePhase.ACTIVE, getLifecyclePhase());
         }
     }
 
@@ -187,7 +187,7 @@ public abstract class EventBot extends TriggeredBot {
             eventBus.publish(new MatcherRegisteredEvent(wonNodeUri));
         } else {
             logger.info("not publishing event for call to onNewAtomCreated() as the bot is not in state {} but {}",
-                    BotLifecyclePhase.ACTIVE, getLifecyclePhase());
+                            BotLifecyclePhase.ACTIVE, getLifecyclePhase());
         }
     }
 
@@ -197,7 +197,7 @@ public abstract class EventBot extends TriggeredBot {
             eventBus.publish(new AtomModifiedEventForMatcher(atomURI));
         } else {
             logger.info("not publishing event for call to onNewAtomCreated() as the bot is not in state {} but {}",
-                    BotLifecyclePhase.ACTIVE, getLifecyclePhase());
+                            BotLifecyclePhase.ACTIVE, getLifecyclePhase());
         }
     }
 
@@ -207,7 +207,7 @@ public abstract class EventBot extends TriggeredBot {
             eventBus.publish(new AtomActivatedEventForMatcher(atomURI));
         } else {
             logger.info("not publishing event for call to onNewAtomCreated() as the bot is not in state {} but {}",
-                    BotLifecyclePhase.ACTIVE, getLifecyclePhase());
+                            BotLifecyclePhase.ACTIVE, getLifecyclePhase());
         }
     }
 
@@ -217,7 +217,7 @@ public abstract class EventBot extends TriggeredBot {
             eventBus.publish(new AtomDeactivatedEventForMatcher(atomURI));
         } else {
             logger.info("not publishing event for call to onNewAtomCreated() as the bot is not in state {} but {}",
-                    BotLifecyclePhase.ACTIVE, getLifecyclePhase());
+                            BotLifecyclePhase.ACTIVE, getLifecyclePhase());
         }
     }
 
@@ -227,7 +227,7 @@ public abstract class EventBot extends TriggeredBot {
             eventBus.publish(new FailureResponseEvent(failedMessageUri, wonMessage));
         } else {
             logger.info("not publishing event for call to onFailureResponse() as the bot is not in state {} but {}",
-                    BotLifecyclePhase.ACTIVE, getLifecyclePhase());
+                            BotLifecyclePhase.ACTIVE, getLifecyclePhase());
         }
     }
 
@@ -237,7 +237,7 @@ public abstract class EventBot extends TriggeredBot {
             eventBus.publish(new SuccessResponseEvent(successfulMessageUri, wonMessage));
         } else {
             logger.info("not publishing event for call to onSuccessResponse() as the bot is not in state {} but {}",
-                    BotLifecyclePhase.ACTIVE, getLifecyclePhase());
+                            BotLifecyclePhase.ACTIVE, getLifecyclePhase());
         }
     }
 
@@ -369,7 +369,7 @@ public abstract class EventBot extends TriggeredBot {
     private EventGeneratingWonMessageSenderWrapper getWonMessageSenderWrapperLazily() {
         if (this.wonMessageSenderWrapper == null) {
             this.wonMessageSenderWrapper = new EventGeneratingWonMessageSenderWrapper(
-                    EventBot.this.getWonMessageSender());
+                            EventBot.this.getWonMessageSender());
         }
         return wonMessageSenderWrapper;
     }
@@ -412,14 +412,14 @@ public abstract class EventBot extends TriggeredBot {
             Event event = null;
             // if the event is connection specific, raise a more specialized event
             switch (type) {
-            case CLOSE:
-            case CONNECT:
-            case CONNECTION_MESSAGE:
-            case OPEN:
-                event = new WonMessageSentOnConnectionEvent(message);
-                break;
-            default:
-                event = new WonMessageSentEvent(message);
+                case CLOSE:
+                case CONNECT:
+                case CONNECTION_MESSAGE:
+                case OPEN:
+                    event = new WonMessageSentOnConnectionEvent(message);
+                    break;
+                default:
+                    event = new WonMessageSentEvent(message);
             }
             getEventBus().publish(event);
         }
